requirements.txt
alabaster==0.7.12
asgiref==3.2.3
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-haystack==2.8.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
DjangoUeditor==2.0
docutils==0.16
idna==2.8
imagesize==1.2.0
itypes==1.1.0
jieba==0.40
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.1
Pillow==6.2.1
Pygments==2.5.2
PyJWT==1.7.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytz==2019.3
redis==3.4.1
requests==2.22.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
Whoosh==2.7.4
